% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_import.R
\name{batch_import}
\alias{batch_import}
\title{Import multiple classroom observations}
\usage{
batch_import(wd = wd(), pattern = NULL, excel.sheet = 1, row.skip = 0)
}
\description{
Import multiple classroom observations
}
