% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stains2018_codes.R
\name{stains2018_codes}
\alias{stains2018_codes}
\title{COPUS codes from: Anatomy of STEM teaching in North American higher-ed}
\usage{
stains2018_codes()
}
\description{
Data set containing COPUS (Classroom Observation for Undergraduate STEM) coding of 2008 STEM classroom observations collected from 548 faculty across 25 institutions in North America.
This function will check to see if the dataset is in your working directory. If not, it will download the file from the repository and then load the dataset into your R environment.
}
\details{
Source: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/48MMF9

Citation: Stains, Marilyne, 2018, "Replication Data for: Anatomy of STEM teaching in North American higher-ed", https://doi.org/10.7910/DVN/48MMF9, Harvard Dataverse, V1
}
